test: move `test_i24_be_bytes_to_i32` test
